PR comments fix variable name move from for loop to stream filtering (and fix logic (was good before)) moved near and gametime commands to utility package
